move a couple things into uiCommon/dl
